rename DocumentObserver to DocumentStore
