+++ packages/houdini-svelte/src/runtime/stores/subscription.ts
@@ -1,4 +1,4 @@
-import type { DocumentObserver } from '$houdini/runtime/client/documentObserver'
+import type { DocumentStore } from '$houdini/runtime/client/documentStore'
 import type { SubscriptionArtifact } from '$houdini/runtime/lib/types'
 import { CompiledSubscriptionKind } from '$houdini/runtime/lib/types'
 import type { GraphQLObject } from 'houdini'
@@ -9,14 +9,14 @@ export class SubscriptionStore<_Data extends GraphQLObject, _Input extends {}> {
 	artifact: SubscriptionArtifact
 	kind = CompiledSubscriptionKind
 
-	private store: DocumentObserver<_Data, _Input>
+	private store: DocumentStore<_Data, _Input>
 
 	constructor({ artifact }: { artifact: SubscriptionArtifact }) {
 		this.artifact = artifact
 		this.store = getClient().observe({ artifact: this.artifact })
 	}
 
-	subscribe(...args: Parameters<DocumentObserver<_Data, _Input>['subscribe']>) {
+	subscribe(...args: Parameters<DocumentStore<_Data, _Input>['subscribe']>) {
 		return this.store?.subscribe(...args)
 	}
