+++ packages/houdini/src/runtime/client/plugins/cache.test.ts
@@ -6,8 +6,8 @@ import { Cache } from '../../cache/cache'
 import { CachePolicy } from '../../lib'
 import { setMockConfig } from '../../lib/config'
 import { ArtifactKind, DataSource } from '../../lib/types'
-import type { ClientPlugin } from './../documentObserver'
-import { DocumentObserver } from './../documentObserver'
+import type { ClientPlugin } from '../documentStore'
+import { DocumentStore } from '../documentStore'
 import { cachePolicyPlugin } from './cache'
 
 /**
@@ -170,7 +170,7 @@ test('CacheOnly', async function () {
 /**
  * Utilities for testing the cache plugin
  */
-export function createStore(plugins: ClientPlugin[]): DocumentObserver<any, any> {
+export function createStore(plugins: ClientPlugin[]): DocumentStore<any, any> {
 	const client = new HoudiniClient({
 		url: 'URL',
 		pipeline() {
@@ -178,7 +178,7 @@ export function createStore(plugins: ClientPlugin[]): DocumentObserver<any, any>
 		},
 	})
 
-	return new DocumentObserver({
+	return new DocumentStore({
 		pipeline: plugins,
 		client,
 		cache: true,
