+++ packages/houdini/src/runtime/client/plugins/injectedPlugins.ts
@@ -1,7 +1,7 @@
 // the actual contents of this file will be overwritten by the runtime generator
 // to include imports from plugins so that they can register middlewares as
 // part of the generation pipeline
-import type { ClientPlugin } from '../documentObserver'
+import type { ClientPlugin } from '../documentStore'
 
 const plugins: ClientPlugin[] = []
