+++ site/src/routes/api/client-plugins/+page.svx
@@ -254,7 +254,7 @@ to your hooks. For a full summary, please refer to the the `ClientPluginEnterHan
 ## Type Definitions
 
 The best source of truth for the type definitions are exported from your `$houdini`
-package. You can see them [here](https://github.com/HoudiniGraphql/houdini/tree/main/packages/houdini/src/runtime/client/documentObserver.ts). They've been summarized below for reference but this copy may be out of date.
+package. You can see them [here](https://github.com/HoudiniGraphql/houdini/tree/main/packages/houdini/src/runtime/client/DocumentStore.ts). They've been summarized below for reference but this copy may be out of date.
 If you find a discrepancy, please let us know on GitHub.
 
 ```typescript
